+++ folio-integration-kong/src/main/java/org/folio/tools/kong/utls/RoutingEntryUtils.java
@@ -1,20 +1,13 @@
 package org.folio.tools.kong.utls;
 
-import static java.util.Arrays.stream;
-import static java.util.Collections.emptyList;
-import static org.apache.commons.collections4.CollectionUtils.containsAny;
-
 import java.util.List;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
-import org.apache.commons.collections4.CollectionUtils;
 import org.folio.common.domain.model.RoutingEntry;
-import org.springframework.http.HttpMethod;
+import org.folio.common.gateway.utils.GatewayRouteUtils;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class RoutingEntryUtils {
-
-  private static final List<String> ALL_METHODS = getAllHttpMethods();
 
   /**
    * Provides a list with http method names for {@link RoutingEntry} object.
@@ -23,17 +16,6 @@ public class RoutingEntryUtils {
    * @return {@link List} with {@link String} method names
    */
   public static List<String> getMethods(RoutingEntry re) {
-    var methods = re.getMethods();
-    if (CollectionUtils.isEmpty(methods)) {
-      return emptyList();
-    }
-
-    return containsAny(methods, "*") ? ALL_METHODS : methods;
-  }
-
-  private static List<String> getAllHttpMethods() {
-    return stream(HttpMethod.values())
-      .map(HttpMethod::name)
-      .toList();
+    return GatewayRouteUtils.getMethods(re);
   }
 }